+++ atlasdb-client/src/main/java/com/palantir/atlasdb/table/description/TableDefinition.java
@@ -179,9 +179,7 @@ public class TableDefinition extends AbstractDefinition {
     }
 
     /**
-     * Prefix the row with a hash of the first N row components
-     *
-     * This is not a feature for tables with range scanning enabled.
+     * Prefix the row with a hash of the first N row components.
      */
     public void hashFirstNRowComponents(int numberOfComponents) {
         Preconditions.checkState(state == State.DEFINING_ROW_NAME,
